+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 1.1.0
+
+- `Option`: added utility functions `mapChoice`, `mapChoice_` and `compareSome`.
+- `Contract`: added functions `dismissContractMessages` and `restoreContractMessages` to allow disabling of specific error message printouts at test time.
+- Improved support for development on Windows.
+- Updated python scripts to work with Python 3.
+
 ## 1.0.0
 Initial commit